requirements.txt
pyodbc-5.2.0
bcrypt-4.3.0